# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Task API",
    "version": "1.0"
  },
  "paths": {
    "/tasks": {
      "get": {
        "summary": "List all tasks",
        "responses": {
          "200": {
            "description": "A list of tasks"
          }
        }
      },
      "post": {
        "summary": "Create a new task",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string" }
                },
                "required": ["title"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Task created"
          },
          "400": {
            "description": "Missing or empty title"
          }
        }
      }
    },
    "/tasks/{id}": {
  "get": {
    "summary": "Get a single task by id",
    "parameters": [
      {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "integer" }
      }
    ],
    "responses": {
      "200": { "description": "The matching task" },
      "404": { "description": "Task not found" }
    }
  },
  "put": {
    "summary": "Update a task's title and/or done status",
    "parameters": [
      {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "integer" }
      }
    ],
    "requestBody": {
      "required": false,
      "content": {
        "application/json": {
          "schema": {
            "type": "object",
            "properties": {
              "title": { "type": "string" },
              "done": { "type": "boolean" }
            }
          }
        }
      }
    },
    "responses": {
      "200": { "description": "Updated task" },
      "400": { "description": "Invalid body" },
      "404": { "description": "Task not found" }
    }
  },
  "delete": {
    "summary": "Delete a task",
    "parameters": [
      {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "integer" }
      }
    ],
    "responses": {
      "204": { "description": "Task deleted" },
      "404": { "description": "Task not found" }
    }
  }
}
  }
}